requirements.txt
fire==0.7.1
fonttools==4.61.1
lxml==6.0.2
numpy==2.4.2
opencv-python-headless==4.13.0.90
pdf2docx==0.5.8
PyMuPDF==1.26.7
python-docx==1.2.0
termcolor==3.3.0
typing_extensions==4.15.0
customtkinter
Pillow
